Convert Student UI to MVVM
