+++ lib/main.dart
@@ -1,5 +1,4 @@
 //import 'package:ez_edu/templates/templateMenu.dart';
-import 'package:ezedu/screens/home/main/widgets/home_main_screen.dart';
 import 'package:ezedu/screens/tutor/login/widgets/tutor_login_main_screen.dart';
 import 'package:ezedu/screens/tutor/profile/widgets/tutor_profile_main_screen.dart';
 import 'package:ezedu/screens/tutor/register/widgets/tutor_register_view.dart';
@@ -21,7 +20,7 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primaryColor: primaryColor,
       ),
-      home: TutorRegistration(),
+      //home: TutorRegistration(),
     );
   }
 }
