+++ lib/screens/student/chat/student_chat_main_screen.dart
@@ -1,0 +1,13 @@
+import 'package:ezedu/screens/student/chat/student_chat_body.dart';
+import 'package:flutter/material.dart';
+
+class StudentChatMainScreen extends StatelessWidget{
+  const StudentChatMainScreen({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: StudentChatBody(),
+    );
+  }
+}
