+++ lib/screens/student/chat/student_chat_viewmodel.dart
@@ -1,0 +1,5 @@
+import 'package:ezedu/screens/viewmodel.dart';
+
+class StudentChatViewModel extends ViewModel{
+  //view model function here
+}
