+++ lib/screens/student/classes/student_classes_viewmodel.dart
@@ -1,0 +1,5 @@
+import 'package:ezedu/screens/viewmodel.dart';
+
+class StudentClassesViewModel extends ViewModel {
+  //view model function here
+}
