+++ lib/screens/student/home/student_home_main_screen.dart
@@ -1,0 +1,15 @@
+
+import 'package:ezedu/screens/student/home/student_home_body.dart';
+import 'package:flutter/material.dart';
+
+class StudentHomeMainScreen extends StatelessWidget {
+  const StudentHomeMainScreen({Key? key}) : super(key: key);
+
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body:  StudentHomeBody(),
+    );
+  }
+}
