+++ lib/screens/student/home/student_home_viewmodel.dart
@@ -1,0 +1,6 @@
+
+import 'package:ezedu/screens/viewmodel.dart';
+
+class StudentHomeViewModel extends ViewModel{
+  // view model function here
+}
