+++ lib/screens/student/notes/student_notes_viewmodel.dart
@@ -1,0 +1,6 @@
+
+import 'package:ezedu/screens/viewmodel.dart';
+
+class StudentNotesViewModel extends ViewModel{
+  // view model function here
+}
