+++ lib/screens/student/payment/student_payment_main_screen.dart
@@ -1,0 +1,13 @@
+import 'package:ezedu/screens/student/payment/student_payment_body.dart';
+import 'package:flutter/material.dart';
+
+class StudentPaymentMainScreen extends StatelessWidget {
+  const StudentPaymentMainScreen({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: StudentPaymentBody(),
+    );
+  }
+}
