+++ lib/screens/student/payment/student_payment_viewmodel.dart
@@ -1,0 +1,5 @@
+import 'package:ezedu/screens/viewmodel.dart';
+
+class StudentPaymentViewModel extends ViewModel {
+  // view model function here
+}
