+++ lib/screens/student/tab/student_tab_viewmodel.dart
@@ -1,0 +1,6 @@
+
+import 'package:ezedu/screens/viewmodel.dart';
+
+class StudentTabViewModel extends ViewModel{
+  // view model function here
+}
